// src/app/(main)/institution/types.ts
export type InstitutionData = {
  _id: string;
  name: string;
  sub_district: string;
  address: string;
  gudep_man: string;
  gudep_woman: string;
  head_gudep_man: string;
  head_gudep_woman: string;
  nta_head_gudep_man: string;
  nta_head_gudep_woman: string;
  headmaster_name: string;
  headmaster_number: string;
};